src/requirements.txt
asgiref==3.2.10
Django==3.0.8
legacy==0.1.6
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
